display chatroom
When a chatroom li element is clicked, the active chatroom is changed.  In the chatroom panel, the chatroom name is displayed as well as the members and messages count.  Messages should also be displayed with the correct classes and styling, but this cannot be tested until the website is able to create and send messages.