signalr createchatroom
the create chatroom method has been changed to an ajax function that calls a jsonresult method so that it can call signalr methods from the client that display the new chatroom on all the connected clients' chatroom lists.  However there are still bugs, and it only displays the chatroom for the person who created the chatroom.  I need to change the way that signalr tries to communicate other clients.