active chatroom
When a new chatroom is created.  The chatroom is displayed and the active chatroom list item is changed to the new chatroom only if the signed in user is the chatroom's creator.  Otherwise the new chatroom is added to the chatroom list but is not set to active and is not displayed in the chatroom panel.  There is still an issue when running two instances of the web application.  When one instance of the web app signs in , if the other instance is already signed in then the current user will be changed to the user that just recently signed in.  This error only occurs when the website posts back, like when the user clicks the manage account link.  However since all of the database calls are ajax functions, the db calls are asynchronous and a full post back does not occur.  This error could be fixed by using containerization which isolates each instance of the app.