join chatroom
users can now search for and join chatrooms.  when typing in the search chatrooms text box, the jquery autocomplete widget is used to display a select list of chatrooms that start with the text in the text box.  Only  public chatrooms that the user is not already a member of are displayed.  when the user clicks on an item in the select list, they are added to the chatroom, the signalr group, and the signalr group.  Afterwards, the chatroom is added to the chatroom list and is displayed in the chatroom panel.