auto scrolling
when a chatroom is displayed, it automatically scrolls all the way to the bottom.  When a new message appears in that chatroom, it automatically scrolls down so the new message is visible.  However, it only does this if it is already at the bottom.  If the user scrolls up to look at past messages and a new message comes in, it does not automatically scroll down and disrupt the user.