Add members popup
Added a bootstrap modal for adding members to a chatroom.  The modal is toggled when the user clicks the invite users action menu item.  Users can click a link which adds an extra text box to the form and has an x button next to it to delete the added text box.