members count and message count
The logic for incrementing the message count has been moved to the signalr "ReceiveMessage" function so that is is not executed repeatedly for every message when initially displaying a chatroom.  Also, the members count is incremented or decremented when a user joins or leaves a chatroom.  However, right now it does not work when a user is added to a chatroom by someone else.  To fix this I just need to ensure that a message is sent to the chatroom when a user is added to a chatroom by someone else.