activeChatroom and getChatroom
a global object was created to store the entire active chatroom instead of just the chatroom id.  Now ajax calls do not need to be made every time the user wants to access certain properties of the active chatroom.
Also, when the getChatroom function is called, the response is returned so that the user can choose if they want to display the chatroom or perform some other action, making the function more reusable.  getChatroom returns a promise so a callback can be attached to it using the then function.